extract all the things
