+++ lib/scarpe/libui/button.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+class Scarpe
+  class << self
+    def button(text)
+      @button = UI.new_button(text)
+
+      UI.button_on_clicked(@button) do
+        puts "herpderpin in the dark"
+        yield
+      end
+
+      # We're appending this to the top level box. (Note, may want to apply to "parent" long term)
+      UI.box_append($vbox, @button, 0)
+    end
+  end
+end
